Use sys.executable when spawning python subprocesses
On Windows, the way executables are looked up when starting a subprocess
differs from POSIX-like systems, where $PATH is consulted.

This may result in running the default system install of Python instead
of the Python interpreter from the virtualenv that MkDocs itself is
running under, causing a mismatch of versions and/or modules, as seen in
https://github.com/zoni/mkdocs-diagrams/issues/2.

Instead of calling "python", try obtaining full path to the current
interpreter using sys.executable and use that for subprocesses so that
we're invoking the same interpreter.